797 Set the map language to UI language in new branch (#6)
This is a solution of https://github.com/photoview/photoview/issues/789.
The language of the map in Place will be set to the same as the UI. If
the UI language is NOT supported by Mapbox, then the English map will be
provided.
This is a new pull request of setting the map language to UI language,
and I commit it to a new branch other than master. The old one will be
closed soon.
As per the comment in the old pull request
https://github.com/photoview/photoview/pull/790, I deleted the settings
of zoom and center and deleted the setting of projection, which means
the map language cannot be changed.
I have looked at the commit setting the projection, it said "chore: add
mapbox globe view", so I guess the committer just wants to show the
mapbox with the whole world map initially. But I think maybe he or she
doesn't know what the projection really means, in fact, global
projection just makes a spherical map. So, I add zoom: 1 to show the
mapbox with the whole world map initially.

---------

Co-authored-by: WindLi001 <lichenggang1@hotmail.com>