Add ability to allow share albums to be downloaded
Shares are currently not downloadable in full, you must download each
image individually. This commit adds the ability to set a share as
downloadable, and it adds a download button to the share.

This commit does not include the generated schemas as it seems like they
were generating incorrectly. They will be in the next commit.